# specs/001-stock-portfolio-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MyStock API
  description: |
    RESTful API for MyStock stock portfolio management application.
    
    ## Features
    - User authentication with JWT tokens
    - Watchlist management with drag-and-drop ordering
    - Portfolio tracking across three categories (장기/단기/정찰병)
    - Real-time stock data integration with Alpha Vantage API
    - Dark mode preference persistence
    
    ## Authentication
    Most endpoints require authentication via JWT bearer token.
    Obtain token via `/api/v1/auth/login` endpoint.
    
  version: 1.0.0
  contact:
    name: MyStock API Support
    email: support@mystock3.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mystock3.example.com/api/v1
    description: Production server
  - url: https://api-staging.mystock3.example.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /auth/signup:
    post:
      tags:
        - auth
      summary: Create new user account
      description: Register a new user with email and password
      operationId: signup
      security: []  # No auth required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Email already registered"
                error_code: "DUPLICATE_EMAIL"
                timestamp: "2025-11-05T10:00:00Z"
  /auth/login:
    post:
      tags:
        - auth
      summary: Authenticate user
      description: Login with email and password, receive JWT token (7-day expiration)
      operationId: login
      security: []  # No auth required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Invalid email or password"
                error_code: "INVALID_CREDENTIALS"
                timestamp: "2025-11-05T10:00:00Z"
  /auth/logout:
    post:
      tags:
        - auth
      summary: Logout user
      description: Invalidate current JWT token (client-side token removal)
      operationId: logout
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'

  # User Endpoints
...
  /watchlist:
    get:
      tags:
        - watchlist
      summary: Get user's watchlist
      description: Retrieve all watchlist items for authenticated user, ordered by display_order
      operationId: getWatchlist
      responses:
        '200':
          description: Watchlist retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WatchlistItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - watchlist
      summary: Add stock to watchlist
      description: Add a new stock to user's watchlist
      operationId: addToWatchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistItemCreate'
      responses:
        '201':
          description: Stock added to watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Stock already in watchlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "이미 관심종목에 추가된 종목입니다"
                error_code: "DUPLICATE_STOCK"
                timestamp: "2025-11-05T10:00:00Z"
  /watchlist/{item_id}:
    get:
      tags:
        - watchlist
      summary: Get watchlist item
      description: Retrieve a specific watchlist item by ID
      operationId: getWatchlistItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: Watchlist item retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - watchlist
      summary: Update watchlist item
      description: Update memo or display order for a watchlist item
      operationId: updateWatchlistItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistItemUpdate'
      responses:
        '200':
          description: Watchlist item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - watchlist
      summary: Remove stock from watchlist
      description: Delete a watchlist item
      operationId: deleteWatchlistItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204':
          description: Watchlist item deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /watchlist/reorder:
    post:
      tags:
        - watchlist
      summary: Reorder watchlist items
      description: Update display order of multiple watchlist items (drag-and-drop)
      operationId: reorderWatchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistReorderRequest'
      responses:
        '200':
          description: Watchlist reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WatchlistItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Portfolio Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint (7-day expiration)
  parameters:
    ItemId:
      name: item_id
      in: path
      description: Watchlist item UUID
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 6
          description: Password (minimum 6 characters)
          example: "password123"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "password123"
    AuthResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - user
      properties:
        access_token:
          type: string
          description: JWT bearer token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          enum: [bearer]
          example: "bearer"
        expires_in:
          type: integer
          description: Token expiration in seconds (7 days = 604800)
          example: 604800
        user:
          $ref: '#/components/schemas/UserResponse'

    # User Schemas
    UserResponse:
      type: object
      required:
        - id
        - email
        - created_at
        - preferences
      properties:
        id:
          type: string
          format: email
          example: "user@example.com"
        email:
          type: string
          format: email
          example: "user@example.com"
        created_at:
          type: string
          format: date-time
          example: "2025-11-05T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-11-05T10:00:00Z"
        preferences:
          $ref: '#/components/schemas/UserPreferences'
    UserPreferences:
      type: object
      properties:
        dark_mode:
          type: boolean
          default: false
          example: true
        language:
          type: string
          default: "ko"
          example: "ko"
...
    WatchlistItemCreate:
      type: object
      required:
        - symbol
      properties:
        symbol:
          type: string
          pattern: '^[A-Z]{1,5}$'
          description: Stock ticker symbol
          example: "AAPL"
        memo:
          type: string
          maxLength: 50
          description: User's note about the stock (max 50 characters)
          example: "매수 타이밍 지켜보는 중"
    WatchlistItemUpdate:
      type: object
      properties:
        memo:
          type: string
          maxLength: 50
          example: "매수 완료"
        display_order:
          type: integer
          minimum: 1
          example: 3
    WatchlistItemResponse:
      type: object
      required:
        - id
        - user_id
        - symbol
        - company_name
        - display_order
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        user_id:
          type: string
          format: email
          example: "user@example.com"
        symbol:
          type: string
          example: "AAPL"
        company_name:
          type: string
          example: "Apple Inc."
        memo:
          type: string
          example: "매수 타이밍 지켜보는 중"
        display_order:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2025-11-05T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-11-05T11:00:00Z"
        current_price:
          type: number
          format: double
          description: Current stock price (from cache/API)
          example: 178.32
        change_percent:
          type: number
          format: double
          description: Price change percentage
          example: 1.22
    WatchlistReorderRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          description: Array of item IDs in new order
          items:
            type: string
            format: uuid
          example:
            - "550e8400-e29b-41d4-a716-446655440003"
            - "550e8400-e29b-41d4-a716-446655440001"
            - "550e8400-e29b-41d4-a716-446655440002"

    # Portfolio Schemas
...
    ErrorResponse:
      type: object
      required:
        - detail
        - error_code
        - timestamp
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"
        error_code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp
          example: "2025-11-05T10:00:00Z"
        field_errors:
          type: object
          description: Field-specific validation errors
          additionalProperties:
            type: string
          example:
            email: "Invalid email format"
            password: "Password must be at least 6 characters"
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Validation error"
            error_code: "VALIDATION_ERROR"
            timestamp: "2025-11-05T10:00:00Z"
            field_errors:
              symbol: "Symbol must be uppercase letters only"
    Unauthorized:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Authentication required"
            error_code: "UNAUTHORIZED"
            timestamp: "2025-11-05T10:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Resource not found"
            error_code: "NOT_FOUND"
            timestamp: "2025-11-05T10:00:00Z"